shader create info

+++ Code/Hawk/Console/ConsoleAPI.cpp
@@ -200,7 +200,7 @@ void ConsoleAPI::WriteLine(const std::string& p_Text, Color p_Color, Color p_BgC
 void ConsoleAPI::Write(const std::string& p_Text, Color p_Color, Color p_BgColor)
 {
 	SetTextColor(p_Color, p_BgColor);
-	THROW_IF_NOT(::WriteConsole(n_hOut, p_Text.c_str(), p_Text.length(), LPDWORD(), nullptr), "Failed to write to console");
+	THROW_IF_NOT(::WriteConsole(n_hOut, p_Text.c_str(), (DWORD)p_Text.length(), LPDWORD(), nullptr), "Failed to write to console");
 }
 
 bool ConsoleAPI::HasNextChar()
